refactor: cleanup utils
