+++ src/components/sources/VectorSource.tsx
@@ -12,7 +12,7 @@ import { type BaseProps } from "../../types/BaseProps";
 import { type FilterExpression } from "../../types/MapLibreRNStyles";
 import { type PressEventWithFeatures } from "../../types/PressEvent";
 import { cloneReactChildrenWithProps, isAndroid } from "../../utils";
-import { getFilter } from "../../utils/filterUtils";
+import { getFilter } from "../../utils/getFilter";
 
 const MLRNModule = NativeModules.MLRNModule;
