+++ src/types/MapLibreRNStyles.ts
@@ -115,7 +115,7 @@ export type ExpressionField =
   | ExpressionField[]
   | { [key: string]: ExpressionField };
 
-export type Expression = readonly [ExpressionName, ...ExpressionField[]];
+export type Expression = [ExpressionName, ...ExpressionField[]];
 
 export type FilterExpression = Expression;
 
@@ -650,7 +650,7 @@ export interface SymbolLayerStyle {
    */
   iconHaloColorTransition?: Transition;
   /**
-       * Distance of halo to the icon outline. 
+       * Distance of halo to the icon outline.
 
 The unit is in pixels only for SDF sprites that were created with a blur radius of 8, multiplied by the display density. I.e., the radius needs to be 16 for `@2x` sprites, etc.
        *
