+++ src/utils/getFilter.ts
@@ -1,6 +1,6 @@
 import { type FilterExpression } from "../types/MapLibreRNStyles";
 
-export function getFilter(filter: FilterExpression | undefined): string[] {
+export function getFilter(filter: FilterExpression | undefined): any[] {
   if (!Array.isArray(filter)) {
     return [];
   }
